\name{wa_sl3611_timestamp}
\docType{data}
\alias{wa_sl3611_timestamp}
\title{Waist Accelerometer data subject sl3611}
\description{
 This data set gives "cnt" counts/minute for subject sl3611.  The first column is "tstmp" combined date time information. 
}
\usage{wa_sl3611_timestamp}
\format{A matix 2 x 1741}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}