\name{wa_tp5788}
\docType{data}
\alias{wa_tp5788}
\title{Waist Accelerometer data subject sj6909}
\description{
This data set gives counts/minute for subject sj6909, "Axis 1". The first two columns are date and time information. 
}
\usage{tp5788_wa}
\format{A matix 3 x 3319}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}