\name{wr_sj6909_ts}
\docType{data}
\alias{wr_sj6909_ts}
\title{Wrist Accelerometer data subject sj6909}
\description{
This data set gives counts/minute for subject sj6909, "Axis 1". The first column is TS combined date time information. 
}
\usage{wr_sj6909_ts}
\format{A matix 2 x 1366}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}