\name{wr_sl3611_timestamp}
\docType{data}
\alias{wr_sl3611_timestamp}
\title{Wrist Accelerometer data subject sl3611}
\description{
 This data set gives "cnt" counts/minute for subject sl3611.  The first column is "tstmp" combined date time information. 
}
\usage{wr_sl3611_timestamp}
\format{A matix 2 x 1741}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}