\name{wr_tp5788_cc}
\docType{data}
\alias{wr_tp5788_cc}
\title{Waist Accelerometer data subject sj6909}
\description{
This data set gives counts/minute for subject sj6909, "Axis 1". The first two columns are date and time information, but labeled "calender" and clock". 
}
\usage{wr_tp5788_cc}
\format{A matix 3 x 2092}
\source{Collected at the Energy Balance Laboratory at Vanderbilt University Medical Center (data owner), using a Actigraph GT1M accelerometer.}
\keyword{datasets}